+++ spec/components/parameters/subscriptionId.ts
@@ -2,10 +2,10 @@ import { ParameterObject } from 'openapi3-ts';
 import { createComponentRef } from '../../../utils/ref';
 
 const subscriptionId: ParameterObject = {
-  name: 'webhookId',
+  name: 'subscriptionId',
   in: 'path',
   required: true,
-  description: 'the webhook subscription identifier',
+  description: 'the subscription identifier',
   schema: {
     type: 'string',
   },
