+++ spec/components/schemas/subscription/partial-subscription.ts
