+++ spec/tags/subscriptions.md
@@ -1,4 +1,4 @@
-Webhook subscriptions allow you to receive events in configured webhooks. The available type of events are **MESSAGE** and **MESSAGE_STATUS**:
+Webhook allow you to receive events in configured webhooks. The available type of events are **MESSAGE** and **MESSAGE_STATUS**:
 
 ## MESSAGE
